feat: unwrap git namespace to flat exports + barrel
